refactor(dialogMap): create DialogMapNode file
